requirements.txt
blinker==1.6.2
click==8.1.5
colorama==0.4.6
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
flask-swagger-ui==4.11.1
fpdf==1.7.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.0.0
PyJWT==2.7.0
pyodbc==4.0.39
pypng==0.20220715.0
python-barcode==0.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
qrcode==7.4.2
six==1.16.0
typing_extensions==4.7.1
Werkzeug==2.3.6
